combine selector classes
