+++ molecularnodes/entities/__init__.py
@@ -1,4 +1,3 @@
 from .ensemble import CellPack, Ensemble, StarFile
-from .molecule.base import Molecule
-from .molecule.selector import Selector
+from .molecule.base import Molecule, MoleculeSelector
 from .trajectory import OXDNA, Trajectory
